refactor: rename translation-cache to board-strings
Only two of its five functions are cache reads; the file is pure
functions over the Board["strings"] field, and "cache" wrongly
suggests stateful memoization. Also hoists the test's strings
fixture into a const, removing two non-null assertions.
