+++ src/features/board/translation/board-strings.ts
